Module re_time_panel::recursive_chunks_per_timeline_subscriber

Structs

ChunkTimelineInfo

Cached information about a chunk in the context of a given timeline.

EntityTimelineChunks

Recursive chunk timeline infos for a given timeline & entity.

PathRecursiveChunksPerTimeline

For each entity & timeline, keeps track of all its chunks and chunks of its children.

PathRecursiveChunksPerTimelineStoreSubscriber 🔒

Functions

add_chunk 🔒

remove_chunk 🔒