re_time_panel
Module time_axis

Structs

TimelineAxis 🔒

A piece-wise linear view of a single timeline.

Functions

collect_candidate_gaps 🔒

Returns None to signal an abort.

collect_gaps_with_granularity 🔒

Returns None to signal an abort.

create_ranges 🔒

Collapse any gaps larger or equals to the given threshold.

gap_size_heuristic 🔒

First determine the threshold for when a gap should be closed. Sometimes, looking at data spanning milliseconds, a single second pause can be an eternity. When looking at data recorded over hours, a few minutes of pause may be nothing. We also don’t want to produce a timeline of only gaps. Finding a perfect heuristic is impossible, but we do our best!