re_time_panel::time_axis
Function gap_size_heuristic

fn gap_size_heuristic(time_type: TimeType, times: &TimeHistogram) -> u64

First determine the threshold for when a gap should be closed. Sometimes, looking at data spanning milliseconds, a single second pause can be an eternity. When looking at data recorded over hours, a few minutes of pause may be nothing. We also don’t want to produce a timeline of only gaps. Finding a perfect heuristic is impossible, but we do our best!