re_time_panel
Module time_panel

Structs

TimePanel

A panel that shows entity names to the left, time on the top.

TimePanelItem

Enums

TimePanelSource

Functions

copy_time_properties_context_menu 🔒

Context menu that shows up when interacting with the streams rect.

help 🔒

help_button 🔒

highlight_timeline_row 🔒

Draw the hovered/selected highlight background for a timeline row.

initialize_time_ranges_ui 🔒

interact_with_streams_rect 🔒

Returns a scroll delta

paint_range_highlight 🔒

paint_time_ranges_gaps 🔒

Visually separate the different time segments

time_marker_ui 🔒

A vertical line that shows the current time.

view_everything 🔒

Find a nice view of everything.