re_time_panel
Module time_ranges_ui

egui uses f32 precision for all screen-space ui coordinates, which makes sense, because that is plenty of precision for things on screen.

In this file we need to use f64 precision because when zoomed in, the screen-space time-ranges of are way outside the screen, leading to precision issues with f32.

Structs

Segment

Constants

GAP_EXPANSION_FRACTION 🔒

How much of the gap use up to expand segments visually to either side?

MAX_GAP 🔒

The ideal gap between time segments.

Functions

gap_width

Sze of the gap between time segments.