Crate re_time_panel

Rerun Time Panel

This crate provides a panel that shows all entities in the store and allows control of time and timelines, as well as all necessary ui elements that make it up.

Modules

data_density_graph 🔒

Show the data density over time for a data stream.

paint_ticks 🔒

recursive_chunks_per_timeline_subscriber 🔒

streams_tree_data 🔒

time_axis 🔒

time_control_ui 🔒

time_panel 🔒

time_ranges_ui 🔒

egui uses f32 precision for all screen-space ui coordinates, which makes sense, because that is plenty of precision for things on screen.

time_selection_ui 🔒

Structs

StreamsTreeData

TimePanel

A panel that shows entity names to the left, time on the top.

Enums

TimePanelSource