Crate re_time_panel

Rerun Time Panel

This crate provides a panel that shows allows to control time & timelines, as well as all necessary ui elements that make it up.

Modules

data_density_graph ğŸ”’

Show the data density over time for a data stream.

paint_ticks ğŸ”’

time_axis ğŸ”’

time_control_ui ğŸ”’

time_ranges_ui ğŸ”’

egui uses f32 precision for all screen-space ui coordinates, which makes sense, because that is plenty of precision for things on screen.

time_selection_ui ğŸ”’

Structs

TimePanel

A panel that shows entity names to the left, time on the top.

TimePanelItem ğŸ”’

Enums

TimePanelSource ğŸ”’

Functions

collapsed_time_marker_and_time ğŸ”’

current_time_ui ğŸ”’

help_button ğŸ”’

highlight_timeline_row ğŸ”’

Draw the hovered/selected highlight background for a timeline row.

initialize_time_ranges_ui ğŸ”’

interact_with_streams_rect ğŸ”’

Returns a scroll delta

paint_range_highlight ğŸ”’

paint_time_ranges_gaps ğŸ”’

Visually separate the different time segments

time_marker_ui ğŸ”’

A vertical line that shows the current time.

view_everything ğŸ”’

Find a nice view of everything.